[Quant][core][gpu][improvement] Refactored implementation for conv2d_cudnn to use packed parameters (Reland PR#73510)
Summary:
This a reland of https://github.com/pytorch/pytorch/pull/73510 --
please look at this PR directly for a summary and test plan.

The only change in this PR is we add the ops to check_forward_backward_compatibility.py
to get around the backwards compatibility issue introduced in the previous PR that
changes the name of the cudnn conv operations. We also renamed cudnnpack_utils.h
to utils.h.

ghstack-source-id: 1722671eb2691675ca8b3b16a9b6e66a7a32fd20
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74220